Fix lack of error handling on file removal failure during updates. (#14)
The code was intending to side-effect on the returned error but didn't
due to the local shadowing.

Impact of bug:

If the an update fails later *and* we fail to remove the temporary
file during cleanup, we would report the original failure rather than
the failure to remove the temporary file (it could be argued this is
actually a good thing).

If an update otherwise succeeds and the deletion of the temporary file
fails, we would silently ignore that error rather than reporting it to
the user or calling process. The result is a left-over encrypted
temporary file.